aum_automation/requirements.txt
# Azure SDK dependencies
azure-identity>=1.15.0
azure-mgmt-compute>=30.0.0
azure-mgmt-network>=25.0.0
azure-mgmt-resource>=23.0.0
azure-mgmt-subscription>=3.1.1
azure-mgmt-resourcegraph>=8.0.0

# Excel handling
openpyxl>=3.1.2

# Configuration
pyyaml>=6.0.1

# Utilities
python-dateutil>=2.8.2
requests>=2.31.0

# Logging enhancements
coloredlogs>=15.0.1
